scripts/agent_learning/requirements.txt
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
autogpt>=0.5.0
pydantic>=2.0.0
requests>=2.31.0